improve naming and init function
