+++ src/utils/constants.js
@@ -11,9 +11,6 @@ export const strokeWeight = 1.5;
 export const pointRadius = 8;
 export const fillOpacity = 0.7;
 
-// Keys for i18n for the policy strength text
-export const policyStrengthI18nKeys = ["", "few", "some", "many"];
-
 // note: this matches the breakpoint in styles.css
 export const MOBILE_BREAKPOINT = 640;
 export const DESKTOP_BREAKPOINT = 1200;
